src/main/java/com/github/brunomndantas/jscrapper/support/parser/collection/attribute/reference/CollectionReferenceBooleanAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.collection.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.collection.attribute.CollectionAttributeParser;
import org.openqa.selenium.WebElement;

public class CollectionReferenceBooleanAttributeParser extends CollectionAttributeParser {

    protected CollectionReferenceBooleanAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Boolean.parseBoolean(element.getAttribute(super.attribute));
    }

}
