src/test/java/com/github/brunomndantas/jscrapper/support/parser/collection/attribute/reference/CollectionReferenceBooleanAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.collection.attribute.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class CollectionReferenceBooleanAttributeParserTest {

    @Test
    public void constructorTest() {
        String attribute = "";
        CollectionReferenceBooleanAttributeParser parser = new CollectionReferenceBooleanAttributeParser(attribute);
        assertEquals(null, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
        CollectionReferenceBooleanAttributeParser parser = new CollectionReferenceBooleanAttributeParser("");

        assertEquals(true, parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return "true";
            }
        }));
    }

}
